Limited the Hikari pool size
The free version of Heroku has a limit of 20 connections per database. If there is no limit on the Hikari pool size, it will open as many connections as it can, thus making it impossible to debug the app from 2 machines with the production database. A limit of 5 connections should be enough for 4 instances of the app running simultaneously.